Splitting out the android version of the fp16 conversion.
