+++ java/src/main/java/ai/onnxruntime/platform/package-info.java
@@ -1,0 +1,11 @@
+/*
+ * Copyright (c) 2023, Oracle and/or its affiliates. All rights reserved.
+ * Licensed under the MIT License.
+ */
+
+/**
+ * A package of platform specific code, used to swap out Java implementations which don't run on Android.
+ *
+ * <p>This is the Java version of the package.
+ */
+package ai.onnxruntime.platform;
